requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-debugtoolbar==0.13.1
Flask-WTF==1.0.0
WTForms==3.0.0
google-generativeai==0.3.1
pdfkit==1.0.0
razorpay==1.3.0
python-dotenv==1.0.0
requests==2.31.0
termcolor==2.2.0
gunicorn==20.1.0
Flask-Session==0.5.0 